create inputStyles.utils, update stories
