+++ packages/atlas/src/components/_inputs/Input/Input.styles.ts
@@ -2,77 +2,9 @@ import { css } from '@emotion/react'
 import styled from '@emotion/styled'
 
 import { Text } from '@/components/Text'
-import { cVar, sizes } from '@/styles'
+import { sizes } from '@/styles'
 
-export type InputSize = 'medium' | 'large'
-
-export const sharedInputStyles = {
-  default: css`
-    /* inherit from Text component font size and color */
-    font-size: inherit;
-    color: inherit;
-
-    /* line height needs to be set for inputs explicitly */
-    line-height: ${sizes(6)};
-    border: none;
-    caret-color: ${cVar('colorBackgroundPrimary')};
-    background-color: ${cVar('colorBackgroundMutedAlpha')};
-    transition: ${cVar('animationTransitionFast')};
-    transition-property: background-color, color, box-shadow;
-    box-shadow: inset 0 -1px 0 0 ${cVar('colorBorder')};
-
-    ::placeholder {
-      font: inherit;
-      color: ${cVar('colorTextMuted')};
-    }
-  `,
-  hover: css`
-    background-color: ${cVar('colorBackgroundAlpha')};
-  `,
-  error: css`
-    box-shadow: inset 0 -1px 0 0 ${cVar('colorBorderError')};
-  `,
-  disabled: css`
-    opacity: 0.5;
-    background-color: ${cVar('colorBackgroundMutedAlpha')};
-  `,
-  focus: css`
-    background-color: ${cVar('colorBackgroundMutedAlpha')};
-    box-shadow: inset 0 -1px 0 0 ${cVar('colorBorderPrimary')};
-  `,
-}
-
-export const horizontalPadding = {
-  medium: sizes(4),
-  large: sizes(5),
-} as const
-
-export const verticalPadding = {
-  medium: sizes(2),
-  large: sizes(3),
-} as const
-
-type NodeWidthProps = {
-  leftNodeWidth?: number
-  rightNodeWidth?: number
-}
-
-const getPadding = (size: InputSize, nodeWidthProps?: NodeWidthProps) => {
-  const paddingLeft = nodeWidthProps?.leftNodeWidth
-    ? `${nodeWidthProps?.leftNodeWidth + sizes(2, true)}px`
-    : horizontalPadding[size]
-
-  const paddingRight = nodeWidthProps?.rightNodeWidth
-    ? `${nodeWidthProps?.rightNodeWidth + sizes(2, true)}px`
-    : horizontalPadding[size]
-
-  return css`
-    padding-top: ${verticalPadding[size]};
-    padding-bottom: ${verticalPadding[size]};
-    padding-left: ${paddingLeft};
-    padding-right: ${paddingRight};
-  `
-}
+import { InputSize, NodeWidthProps, getInputPseudoSelectorStyles, getPadding, horizontalPadding } from '../inputs.utils'
 
 type TextInputProps = {
   error?: boolean
@@ -93,23 +25,9 @@ export const TextInput = styled.input<TextInputProps>`
     appearance: textfield;
   }
 
-  ${sharedInputStyles.default};
-
   ${({ inputSize, leftNodeWidth, rightNodeWidth }) => getPadding(inputSize, { leftNodeWidth, rightNodeWidth })};
 
-  :hover:not(:disabled) {
-    ${sharedInputStyles.hover};
-  }
-
-  :focus:not(:disabled) {
-    ${sharedInputStyles.focus};
-  }
-
-  :disabled {
-    ${sharedInputStyles.disabled};
-  }
-
-  ${({ error }) => error && sharedInputStyles.error};
+  ${({ error }) => getInputPseudoSelectorStyles({ error })};
 `
 
 export const InputContainer = styled(Text)`
